A function to calculate area between two profiles

This function takes two profiles defined as (l1,z1) and (l2,z2) as inputs and calculates area between them

area_between(s1, s2 = NA, h = NA, sigma_z = c(NA, NA),
  sigma_l = c(NA, NA))

Arguments

s1
tibble with columns l and z describing first profile

s2
tibble with columns l and z describing second profile

h
if provided by user, the second profile is supposed to be horizontal, with constant height=h (defaults to NA)

sigma_z
a vector of length 1 or 2 providing an estimate of the error in measures of height z1 and z2

sigma_l
a vector of length 1 or 2 providing an estimate of the error in measures of longitudinal coordinates l1 and l2z1 and z2

Value

area

area_by_type

sigma_area

data

Examples

s1 <- tibble(l=c(1,3,5,6,9),
             z=c(1,2,3,2.5,5))
s2 <- tibble(l=c(0.5,2.5,4,6,8),
             z=c(3,1,2,4,3))
area_between(s1,s2)

#> $data
#> # A tibble: 26 x 15
#>        l    z1 p        z2      a type  l_obs z_obs series z1min z1max z2min
#>    <dbl> <dbl> <chr> <dbl>  <dbl> <chr> <dbl> <dbl> <chr>  <dbl> <dbl> <dbl>
#>  1   0.5 NA    obse~  3    NA     NA      0.5     3 s2        NA    NA    NA
#>  2   1    1    obse~  2.5   0     NA      1       1 s1        NA    NA    NA
#>  3   1    1    inte~  2.5  -0.75  lower  NA      NA s2        NA    NA    NA
#>  4   2    1.5  inte~  1.5  NA     lower  NA      NA NA        NA    NA    NA
#>  5   2    1.5  inte~  1.5   0.188 upper  NA      NA NA        NA    NA    NA
#>  6   2.5  1.75 inte~  1     0     NA     NA      NA s1        NA    NA    NA
#>  7   2.5  1.75 obse~  1     0.354 upper   2.5     1 s2        NA    NA    NA
#>  8   3    2    obse~  1.33  0     NA      3       2 s1        NA    NA    NA
#>  9   3    2    inte~  1.33  0.583 upper  NA      NA s2        NA    NA    NA
#> 10   4    2.5  inte~  2     0     NA     NA      NA s1        NA    NA    NA
#> # ... with 16 more rows, and 3 more variables: z2max <dbl>, lmin <dbl>,
#> #   lmax <dbl>
#> 
#> $area
#> [1] -0.4583333
#> 
#> $area_by_type
#> # A tibble: 2 x 2
#>   type   area
#>   <chr> <dbl>
#> 1 lower -2.34
#> 2 upper  1.89
#> 
#> $sigma_area
#> [1] 0
#> 

area_between(s1,h=2)

#> $data
#> # A tibble: 16 x 15
#>        l    z1 p        z2      a type  l_obs z_obs series z1min z1max z2min
#>    <dbl> <dbl> <chr> <dbl>  <dbl> <chr> <dbl> <dbl> <chr>  <dbl> <dbl> <dbl>
#>  1     1   1   obse~     2   0    NA        1   1   s1        NA    NA    NA
#>  2     1   1   obse~     2  -1    lower     1   2   s2        NA    NA    NA
#>  3     3   2   inte~     2  NA    NA       NA  NA   NA        NA    NA    NA
#>  4     3   2   inte~     2  NA    lower    NA  NA   NA        NA    NA    NA
#>  5     3   2   obse~     2   0    NA        3   2   s1        NA    NA    NA
#>  6     3   2   inte~     2   0    NA       NA  NA   s2        NA    NA    NA
#>  7     3   2   inte~     2   0    NA       NA  NA   NA        NA    NA    NA
#>  8     3   2   inte~     2   1    upper    NA  NA   NA        NA    NA    NA
#>  9     5   3   obse~     2   0    NA        5   3   s1        NA    NA    NA
#> 10     5   3   inte~     2   0.75 upper    NA  NA   s2        NA    NA    NA
#> 11     6   2.5 obse~     2   0    NA        6   2.5 s1        NA    NA    NA
#> 12     6   2.5 inte~     2   5.25 upper    NA  NA   s2        NA    NA    NA
#> 13     9   5   obse~     2   0    upper     9   5   s1        NA    NA    NA
#> 14     9   5   obse~     2 NaN    NA        9   2   s2        NA    NA    NA
#> 15   NaN NaN   inte~   NaN  NA    NA       NA  NA   NA        NA    NA    NA
#> 16   NaN NaN   inte~   NaN  NA    NA       NA  NA   NA        NA    NA    NA
#> # ... with 3 more variables: z2max <dbl>, lmin <dbl>, lmax <dbl>
#> 
#> $area
#> [1] 6
#> 
#> $area_by_type
#> # A tibble: 2 x 2
#>   type   area
#>   <chr> <dbl>
#> 1 lower    -1
#> 2 upper     7
#> 
#> $sigma_area
#> [1] 0
#>